fin TP4 (corrigé)

+++ src/main/java/fr/insa/TPCompo/Person.java
@@ -9,7 +9,7 @@ public class Person {
 
     @Id
     @GeneratedValue
-    private int id;
+    private Long id;
     private String name;
     private String surname;
     private String phone;
@@ -18,7 +18,7 @@ public class Person {
     public Person() {
     };
 
-    public Person(int id, String name, String surname, String phone, String city) {
+    public Person(Long id, String name, String surname, String phone, String city) {
         super();
         this.id = id;
         this.name = name;
@@ -35,11 +35,11 @@ public class Person {
         this.city = city;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
